pass isDisabled for the service
