+++ bigbluebutton-html5/imports/ui/components/video-dock/video-menu/container.jsx
@@ -1,14 +1,8 @@
 import React from 'react';
 import { withTracker } from 'meteor/react-meteor-data';
 import { defineMessages, injectIntl } from 'react-intl';
-import Settings from '/imports/ui/services/settings';
-import mapUser from '/imports/ui/services/user/mapUser';
-import Auth from '/imports/ui/services/auth';
-import Meetings from '/imports/api/meetings/';
-import Users from '/imports/api/users/';
 import JoinVideoOptions from './component';
 import VideoMenuService from './service';
-import VideoService from '../service';
 
 const intlMessages = defineMessages({
   joinVideo: {
@@ -29,30 +23,17 @@ const intlMessages = defineMessages({
   },
 });
 
-const JoinVideoOptionsContainer = props => <JoinVideoOptions {...props} />;
+const JoinVideoOptionsContainer = (props) => {
+  const {
+    isSharingVideo,
+    isDisabled,
+    handleJoinVideo,
+    handleCloseVideo,
+    baseName,
+    intl,
+    ...restProps
+  } = props;
 
-export default injectIntl(withTracker(({ intl, handleJoinVideo, handleCloseVideo }) => {
-  const isSharingVideo = VideoMenuService.isSharingVideo();
-  const isWaitingResponse = VideoService.isWaitingResponse();
-  const isConnected = VideoService.isConnected();
-
-  const videoSettings = Settings.video;
-  const enableShare = !videoSettings.viewParticipantsWebcams;
-  const meeting = Meetings.findOne({ meetingId: Auth.meetingID });
-  const LockCam = meeting.lockSettingsProp ? meeting.lockSettingsProp.disableCam : false;
-  const webcamOnlyModerator = meeting.usersProp.webcamsOnlyForModerator;
-
-  const user = Users.findOne({ userId: Auth.userID });
-  const userLocked = mapUser(user).isLocked;
-
-  const isConecting = (!isSharingVideo && isConnected);
-  const isLocked = (LockCam && userLocked) || webcamOnlyModerator;
-  const isDisabled = isLocked
-      || isWaitingResponse
-      || isConecting
-      || enableShare;
-
-  const baseName = Meteor.settings.public.app.basename;
   const videoItems = [
     {
       iconPath: `${baseName}/resources/images/video-menu/icon-swap.svg`,
@@ -70,8 +51,18 @@ export default injectIntl(withTracker(({ intl, handleJoinVideo, handleCloseVideo
     },
   ];
 
-  return {
-    isSharingVideo,
-    videoItems,
-  };
-})(JoinVideoOptionsContainer));
+  return <JoinVideoOptions {...{ videoItems, isSharingVideo, ...restProps }} />;
+};
+
+export default injectIntl(withTracker(({
+  intl,
+  handleJoinVideo,
+  handleCloseVideo,
+}) => ({
+  baseName: VideoMenuService.baseName,
+  isSharingVideo: VideoMenuService.isSharingVideo(),
+  isDisabled: VideoMenuService.isDisabled(),
+  handleJoinVideo,
+  handleCloseVideo,
+  intl,
+}))(JoinVideoOptionsContainer));
